Create Plus Minus (with time + space)
The time complexity of the plusMinus function is O(n), where n represents the size of the input array arr. This is because the function traverses through the array once to tabulate the counts of positive, negative, and zero elements, executing a fixed number of operations for each element along the way.

In the space complexity, the plusMinus function operates at O(1), independent of the size of the input array. It achieves this by utilizing a constant number of integer variables (positiveCount, negativeCount, zeroCount), thereby ensuring a steady memory footprint. Additionally, it refrains from utilizing any supplementary data structures that could grow in proportion to the input size.